src/database.py
import os
import mysql.connector
from dotenv import load_dotenv

load_dotenv()
def get_db_connection():

    try:
        DB_HOST= os.getenv("DB_HOST")
        DB_USER= os.getenv("DB_USER")
        DB_PASSWORD= os.getenv('DB_PASSWORD')
        DB_NAME= os.getenv('DB_NAME')

        connection= mysql.connector.connect(
            host= DB_HOST,
            user= DB_USER,
            password= DB_PASSWORD,
            database= DB_NAME
        )
        return connection
    except mysql.connector.Error as err:
        print(f" error connecting to mysql: {err}")
        return None
